requirements.txt
flask
flask-cors
google-genai
joblib
numpy
pandas
scikit-learn
gunicorn